+++ R/data_copy.R
@@ -145,7 +145,7 @@ spark_data_copy <- function(
   # Many of these issues are marked as resolved, but it appears this is
   # a common regression in Spark and the handling is not uniform across
   # the Spark API.
-  names(df) <- spark_sanitize_names(names(df))
+  names(df) <- spark_sanitize_names(names(df), sc$config)
 
   columns <- lapply(df, function(e) {
     if (is.factor(e))