+++ java/spark-1.5.2/backend.scala
@@ -46,7 +46,7 @@ package sparklyr
 
 class Backend() {
   import java.io.{DataInputStream, DataOutputStream}
-  import java.io.{File, FileOutputStream, IOException}
+  import java.io.{File, FileOutputStream, IOException, FileWriter}
   import java.net.{InetAddress, InetSocketAddress, ServerSocket, Socket}
   import java.util.concurrent.TimeUnit
 
@@ -59,6 +59,7 @@ class Backend() {
   private[this] var isService: Boolean = false
   private[this] var isRemote: Boolean = false
   private[this] var isWorker: Boolean = false
+  private[this] var isBatch: Boolean = false
 
   private[this] var hostContext: String = null
 
@@ -70,6 +71,7 @@ class Backend() {
   private[this] var port: Int = 0
   private[this] var sessionId: Int = 0
   private[this] var connectionTimeout: Int = 60
+  private[this] var batchFile: String = ""
 
   private[this] var sc: SparkContext = null
   private[this] var hc: HiveContext = null
@@ -92,14 +94,6 @@ class Backend() {
     val GetPorts, RegisterInstance, UnregisterInstance = Value
   }
 
-  def getOrCreateHiveContext(sc: SparkContext): HiveContext = {
-    if (hc == null) {
-      hc = new HiveContext(sc)
-    }
-
-    hc
-  }
-
   def getSparkContext(): SparkContext = {
     sc
   }
@@ -114,10 +108,12 @@ class Backend() {
 
   def setType(isServiceParam: Boolean,
               isRemoteParam: Boolean,
-              isWorkerParam: Boolean) = {
+              isWorkerParam: Boolean,
+              isBatchParam: Boolean) = {
     isService = isServiceParam
     isRemote = isRemoteParam
     isWorker = isWorkerParam
+    isBatch = isBatchParam
   }
 
   def setHostContext(hostContextParam: String) = {
@@ -127,10 +123,18 @@ class Backend() {
   def init(portParam: Int,
            sessionIdParam: Int,
            connectionTimeoutParam: Int): Unit = {
+      init(portParam, sessionIdParam, connectionTimeoutParam, "")
+  }
+
+  def init(portParam: Int,
+           sessionIdParam: Int,
+           connectionTimeoutParam: Int,
+           batchFilePath: String): Unit = {
 
     port = portParam
     sessionId = sessionIdParam
     connectionTimeout = connectionTimeoutParam
+    batchFile = batchFilePath
 
     logger = new Logger("Session", sessionId)
 
@@ -193,8 +197,85 @@ class Backend() {
     if (!isService) System.exit(0)
   }
 
+  def batch(): Unit = {
+    new Thread("starting batch rscript thread") {
+      override def run(): Unit = {
+        try {
+          logger.log("is starting batch rscript")
+
+          val rscript = new Rscript(logger)
+
+          val sparklyrGateway = "sparklyr://localhost:" + port.toString() + "/" + sessionId
+          logger.log("will be using rscript gateway: " + sparklyrGateway)
+
+          var sourceFile: File = new java.io.File("sparklyr-batch.R")
+          if (!sourceFile.exists) {
+            logger.log("tried to find source under working folder: " + (new File(".").getAbsolutePath()))
+            logger.log("tried to find source under working files: " + (new File(".")).listFiles.mkString(","))
+
+            sourceFile = new File(rscript.getScratchDir() + File.separator + "sparklyr-batch.R")
+            if (!sourceFile.exists) {
+
+              logger.log("tried to find source under scratch folder: " + rscript.getScratchDir().getAbsolutePath())
+              logger.log("tried to find source under scratch files: " + rscript.getScratchDir().listFiles.mkString(","))
+
+              sourceFile = new File(batchFile)
+            }
+          }
+
+          val sourceLines = scala.io.Source.fromFile(sourceFile).getLines
+
+          val modifiedFile: File = new File(rscript.getScratchDir() + File.separator + "sparklyr-batch-mod.R")
+          val outStream: FileWriter = new FileWriter(modifiedFile)
+          outStream.write("options(sparklyr.connect.master = \"" + sparklyrGateway + "\")")
+          outStream.write("\n\n");
+          for (line <- sourceLines) {
+            outStream.write(line + "\n")
+          }
+          outStream.flush()
+
+          logger.log("wrote modified batch rscript: " + modifiedFile.getAbsolutePath())
+
+          val customEnv: Map[String, String] = Map()
+          val options: Map[String, String] = Map()
+
+          rscript.init(
+            List(),
+            modifiedFile.getAbsolutePath(),
+            customEnv,
+            options
+          )
+        } catch {
+          case e: java.lang.reflect.InvocationTargetException =>
+            e.getCause() match {
+              case cause: Exception => {
+                logger.logError("failed to invoke batch rscript: ", cause)
+                System.exit(1)
+              }
+              case _ => {
+                logger.logError("failed to invoke batch rscript: ", e)
+                System.exit(1)
+              }
+            }
+          case e: Exception => {
+            logger.logError("failed to run batch rscript: ", e)
+            System.exit(1)
+          }
+        }
+      }
+    }.start()
+  }
+
   def run(): Unit = {
     try {
+
+      if (isBatch) {
+        // spark context needs to be created for spark.files to be accessible
+        org.apache.spark.SparkContext.getOrCreate()
+
+        batch()
+      }
+
       initMonitor()
       while(isRunning) {
         bind()