+++ java/spark-1.5.2/shell.scala
@@ -8,11 +8,12 @@ object Shell {
   def main(args: Array[String]): Unit = {
     if (args.length > 4 || args.length < 2) {
       System.err.println(
-        "Usage: Backend port id [--service] [--remote]\n" +
+        "Usage: Backend port id [--service] [--remote] [--batch file.R]\n" +
         "  port:      port the gateway will listen to\n" +
         "  id:        arbitrary numeric identifier for this backend session\n" +
         "  --service: prevents closing the connection from closing the backen\n" +
-        "  --remote:  allows the gateway to accept remote connections\n"
+        "  --remote:  allows the gateway to accept remote connections\n" +
+        "  --batch:   local path to R file to be executed in batch mode\n"
       )
 
       System.exit(-1)
@@ -27,10 +28,14 @@ object Shell {
 
     val isService = args.contains("--service")
     val isRemote = args.contains("--remote")
+    val isBatch = args.contains("--batch")
+
+    var batchFile = ""
+    if (isBatch) batchFile = args(args.indexOf("--batch") + 1)
 
     backend = new Backend()
-    backend.setType(isService, isRemote, false)
-    backend.init(port, sessionId, connectionTimeout)
+    backend.setType(isService, isRemote, false, isBatch)
+    backend.init(port, sessionId, connectionTimeout, batchFile)
   }
 
   def getBackend(): Backend = {